requirements.txt
streamlit
google-cloud-aiplatform
google-auth
PyPDF2